Fix bug in binomial heap when merging roots with equal priority
When merging two trees whose roots have the same priority, it was not ensured
that the resulting root is the current minimum (_min). As a result, _min could
point to a non-root node, which caused an infinite loop in BinomialHeap::unlace
when trying to pop the minimum. A new test case in heap_test.cc demonstrates
this issue.